refactor the different classes into their own packages
