+++ app/src/androidTest/java/io/github/vinge1718/restaurant/MainActivityInstrumentationTest.java
@@ -5,6 +5,8 @@ import android.support.test.rule.ActivityTestRule;
 import org.junit.Rule;
 import org.junit.Test;
 
+import io.github.vinge1718.restaurant.ui.MainActivity;
+
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.action.ViewActions.closeSoftKeyboard;
