+++ app/src/androidTest/java/io/github/vinge1718/restaurant/RestaurantsActivityInstrumentationTest.java
@@ -6,6 +6,8 @@ import android.view.View;
 import org.junit.Rule;
 import org.junit.Test;
 
+import io.github.vinge1718.restaurant.ui.RestaurantsActivity;
+
 import static android.support.test.espresso.Espresso.onData;
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
