+++ app/src/main/java/io/github/vinge1718/restaurant/models/Restaurant.java
@@ -1,4 +1,4 @@
-package io.github.vinge1718.restaurant;
+package io.github.vinge1718.restaurant.models;
 
 import java.util.ArrayList;
