+++ app/src/main/java/io/github/vinge1718/restaurant/ui/RestaurantsActivity.java
@@ -1,21 +1,21 @@
-package io.github.vinge1718.restaurant;
+package io.github.vinge1718.restaurant.ui;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.View;
-import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import java.io.IOException;
 import java.util.ArrayList;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import io.github.vinge1718.restaurant.R;
+import io.github.vinge1718.restaurant.models.Restaurant;
+import io.github.vinge1718.restaurant.services.YelpService;
 import okhttp3.Call;
 import okhttp3.Callback;
 import okhttp3.Response;
