+++ app/src/test/java/io/github/vinge1718/restaurant/RestaurantsActivityTest.java
@@ -8,6 +8,8 @@ import org.junit.runner.RunWith;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
 
+import io.github.vinge1718.restaurant.ui.RestaurantsActivity;
+
 import static junit.framework.TestCase.assertEquals;
 import static junit.framework.TestCase.assertNotNull;
